backend/requirements.txt
# Phase II - Full-Stack Todo Application
# Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database & ORM
sqlmodel==0.0.14
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dotenv==1.0.0
